+++ src/main/java/luis/josh/catan/host/game/actions/PlaceSettlement.java
@@ -24,11 +24,11 @@ public class PlaceSettlement implements Action{
 
     @Override
     public JSONObject[] execute(JSONObject data, Player player) {
-        int start = (int)(long)data.get("start");
+        int start = (int)data.get("start");
         JSONObject location = (JSONObject)data.get("tile");
-        int row = (int)(long)location.get("row");
-        int col = (int)(long)location.get("col");
-        int vertex = (int)(long)location.get("vertex");
+        int row = (int)location.get("row");
+        int col = (int)location.get("col");
+        int vertex = (int)location.get("vertex");
 
         Tile tile = board.tiles[row][col];
         if(tile == null) {
